Organize Ilmomasiina components by route
